added autocmds with filetype detection
Over the years, I've used many languages and so these settings are
purely based on what settings I liked for those languages. This might
not suit everybody, but should be easy enough to customize according
to taste. A brief :help on that setting should be of help or look up
the details of that filetype plugin.